+++ trixnity-client/src/commonMain/kotlin/de/connect2x/trixnity/client/media/InsufficientSpaceException.kt
@@ -1,0 +1,6 @@
+package de.connect2x.trixnity.client.media
+
+class InsufficientSpaceException(
+    val fileSize: Long,
+    val availableSpace: Long,
+) : IllegalStateException("Available space $availableSpace bytes insufficient for file with size $fileSize bytes")