+++ trixnity-clientserverapi/trixnity-clientserverapi-client/src/commonMain/kotlin/de/connect2x/trixnity/clientserverapi/client/DownloadLimitExceededException.kt
@@ -8,7 +8,10 @@ class DownloadLimitExceededException(
     val maxSize: Long,
     message: String? = null,
     cause: Throwable? = null
-) : IllegalStateException("File could not be downloaded because it would exceed the limit of $maxSize bytes"),
+) : IllegalStateException(
+    message ?: "File could not be downloaded because it would exceed the limit of $maxSize bytes",
+    cause
+),
     CopyableThrowable<DownloadLimitExceededException> {
     override fun createCopy(): DownloadLimitExceededException {
         return DownloadLimitExceededException(maxSize, message, cause)