+++ CHANGELOG.md
@@ -8,12 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- `expectedSize` parameter to `MediaService` download functions
 
 ### Changed
 
 - bump c2x-settings-plugin version
 
 ### Deprecated
+- `MediaService`: `getMedia`, `getEncryptedMedia` and `getThumbnail` without required `expectedSize` parameter
 
 ### Removed